Extracted importer and extensions rendering into Import and Extensions controllers Next step would be to move functional logic from Admin Controller Later the controller instances in the Settings Controller would be moved elsewhere as well
Created a base Controller class to handle all repeatable variable logic (version, file, directories etc)
